Complete test to run our command-line.
This moves most of the code from `cty.hs` to `Curiosity/Run.hs`, which
makes it possible to call it from `CuriositySpec.hs`.
